Moved scripts/copyAssets.js to resources folder and fixed formatting
